refactor(dead-code): wave 3 — bonifica cross-linguaggio + rimozione cap-lints (clippy torna a mordere)
SCOPERTA STRUTTURALE: .cargo/config.toml conteneva --cap-lints allow dal
bootstrap del monorepo: cappava TUTTI i lint (anche dei crate workspace),
rendendo cieco il gate clippy -D warnings di verify.sh e nascondendo 170
warning di dead code + 117 warning clippy reali. Rimosso dopo la bonifica:
ora il workspace compila a ZERO warning con i lint pieni.

Eliminazioni (triage adversariale multi-agente: catalogo tool DB per i
dispatch per nome, grep cross-crate dei pub, next/dynamic, pattern framework,
git log -S; poi applicazione e validazione con i tre compilatori):
- ~40 dipendenze Cargo in 13 crate (cargo-machete; unico falso positivo
  prost/mcp-proto ripristinato con metadata ignored)
- 148 item Rust + cascate; 27 expect(dead_code, reason) per i contratti
  legittimi (auto-scadono), con cfg_attr(not(test)) dove i test usano l item
- 22 file TS (mcp-connectors 28KB sostituito da PluginManager, pannello
  knowledge legacy pre-wiki, environment-panel, lib/admin-hooks, ...) + 81
  export; knip.json con entrypoint Playwright
- 33 item Python + tool_translator.py intero (layer dialetti mai cablato) +
  blocco PR-4 reasoning_bank + test_classifier_chain.py (testava una classe
  rimossa dal commit 627197d); whitelist vulture per i falsi positivi
- 54 fix clippy meccanici (incl. 5 regex-in-loop -> LazyLock, anticipo del
  refactor tech-debt-rust) + lint stilistici storici in [workspace.lints]
  con allow espliciti motivati (24 crate collegati)

BUG REALI emersi dai warning (non erano dead code):
- run_tests: tool esposto al modello e whitelistato (mig 0218/0286) ma senza
  braccio nel dispatcher — ogni invocazione falliva (task aperto, expect
  documentato sul progenitore)
- route profili-MCP MAI montate: /admin/profiles chiamava
  GET/PUT /api/admin/profiles/:id/mcp-servers e /api/admin/global-mcp-servers
  senza che esistessero ne le route mcp-core ne i proxy Next: montate + proxy
- replace(I18n,I18n) no-op, rami if duplicati fusi (brain_agent_client,
  connection.rs), worker autofix zombie (gia in wave 2)
- mig 0409: cleanup settings orfane dei lettori rimossi
  (nexus_active_routing_pct, routing.intent_deterministic_*) — scovate dal
  gate audit-settings stesso (il ratchet funziona)

Tooling duraturo: scripts/dead-code-report.sh (compilatore senza cap +
machete + knip + vulture) con baseline ratchet .dead-code-baseline.json;
inventario in docs/tech-debt-dead-code.md. I 9 test brain rotti da prima
(mai eseguiti in CI) sono tracciati in task separato.

Saldo: -682 righe brain, ~150 file toccati, pnpm verify verde incluso il
nuovo gate audit-settings e clippy non cappato.

+++ crates/nexus-orchestrator/src/workers/session_persistence.rs
@@ -110,7 +110,7 @@ impl LearningWorker for SessionPersistenceWorker {
         // Statistiche swarm (se presenti)
         let (swarm_tasks_total, swarm_tasks_success) =
             if let Some(swarm) = &context.swarm_result {
-                (swarm.task_results.len(), swarm.success_count as usize)
+                (swarm.task_results.len(), swarm.success_count)
             } else {
                 (0, 0)
             };
